+++ src/app/components/space-cleaning-reports/data-mobile-item.js
@@ -1,0 +1,17 @@
+/**
+ * Report entry in mobile mode
+ * 
+ * @module components/space-cleaning-reports/data-mobile-item
+ * @author Jakub Liput
+ * @copyright (C) 2018-2019 ACK CYFRONET AGH
+ * @license This software is released under the MIT license cited in 'LICENSE.txt'.
+ */
+
+import DataItemBase from './-data-item-base';
+
+export default DataItemBase.extend({
+  tagName: '',
+  classNames: ['data-row'],
+
+  listItem: undefined,
+});