+++ src/app/locales/en/components/space-file-popularity-configuration.js
@@ -1,0 +1,11 @@
+export default {
+  configurationHeader: 'Configuration',
+  configurationText: 'The file-popularity mechanism creates a view index which can be queried to fetch the least popular files. The following parameters are used to configure the popularity function, which is used to sort the view. Detailed information on the popularity function can be found in',
+  configurationTextDoc: 'documentation',
+  lastOpenHourWeight: 'Last open hour weight ',
+  lastOpenHourWeightHint: 'Weight of lastOpenHour parameter taken to calculate the value of the popularity function.',
+  avgOpenCountPerDayWeight: 'Average open count per day',
+  avgOpenCountPerDayWeightHint: 'Weight of avgOpenCountPerDayWeight parameter taken to calculate the value of of the popularity function.',
+  maxAvgOpenCountPerDay: 'Max average count per day',
+  maxAvgOpenCountPerDayHint: 'Maximal value of average open count per day taken to calculate the value of popularity function.',
+};