+++ src/tests/integration/components/space-cleaning-reports-test.js
@@ -3,6 +3,8 @@ import { describe, it, beforeEach } from 'mocha';
 import { setupComponentTest } from 'ember-mocha';
 import hbs from 'htmlbars-inline-precompile';
 import moment from 'moment';
+import { resolve } from 'rsvp';
+import wait from 'ember-test-helpers/wait';
 import { click } from 'ember-native-dom-helpers';
 
 const START_END_TIME_FORMAT = 'D MMM YYYY H:mm:ss';
@@ -33,70 +35,94 @@ describe('Integration | Component | space cleaning reports', function () {
       addEventListener() {},
       removeEventListener() {},
     });
+    this.on('nop', () => {});
   });
 
   it('renders reports', function () {
     const data = this.get('data');
-    this.render(hbs `{{space-cleaning-reports data=data _window=_window}}`);
+    this.on('fetchReports', () => resolve(data));
+    this.render(hbs `<div class="col-content">
+      {{space-cleaning-reports
+        spaceId="space_id1"
+        isCleanEnabled=true
+        fetchReports=(action "fetchReports")
+        _window=_window
+      }}
+    </div>`);
 
-    expect(this.$('tbody tr')).to.have.length(2);
-    const cells = this.$('tbody tr:first-child td');
-    const cellsValues = [
-      moment(data[0].startedAt).format(START_END_TIME_FORMAT),
-      moment(data[0].stoppedAt).format(START_END_TIME_FORMAT),
-      '1 MiB (out of 2 MiB)',
-      String(data[0].filesNumber),
-    ];
-    cellsValues.forEach((value, index) =>
-      expect(cells.eq(index).text().trim()).to.be.equal(value)
-    );
-    expect(cells.eq(4).find('.oneicon.oneicon-checkbox-filled-x')).to.exist;
-    expect(this.$(
-      'tbody tr:last-child td:last-child .oneicon.oneicon-checkbox-filled'
-    )).to.exist;
+    return wait().then(() => {
+      expect(this.$('tbody tr.data-row')).to.have.length(2);
+      const cells = this.$('tbody tr.data-row:first td');
+      const cellsValues = [
+        moment(data[0].startedAt).format(START_END_TIME_FORMAT),
+        moment(data[0].stoppedAt).format(START_END_TIME_FORMAT),
+        '1 MiB (out of 2 MiB)',
+        String(data[0].filesNumber),
+      ];
+      cellsValues.forEach((value, index) =>
+        expect(cells.eq(index).text().trim()).to.be.equal(value)
+      );
+      expect(cells.eq(4).find('.oneicon.oneicon-checkbox-filled-x')).to.exist;
+      expect(this.$(
+        'tbody tr:last-child td:last-child .oneicon.oneicon-checkbox-filled'
+      )).to.exist;
+    });
   });
 
   it('renders reports in mobile view', function () {
     this.set('_window.innerWidth', 500);
     const data = this.get('data');
-    this.render(hbs `{{space-cleaning-reports data=data _window=_window}}`);
+    this.on('fetchReports', () => resolve(data));
+    this.render(hbs `<div class="col-content">
+      {{space-cleaning-reports
+        spaceId="space_id1"
+        isCleanEnabled=true
+        fetchReports=(action "fetchReports")
+        _window=_window
+      }}
+    </div>`);
 
-    expect(this.$('.one-collapsible-list li')).to.have.length(2);
-    const cells = this.$(
-      '.one-collapsible-list-item:last-child .item-table .content-row .one-label'
-    );
-    const cellsValues = [
-      moment(data[1].startedAt).format(START_END_TIME_FORMAT),
-      moment(data[1].stoppedAt).format(START_END_TIME_FORMAT),
-      '999 KiB (out of 999 KiB)',
-      String(data[1].filesNumber),
-    ];
-    cellsValues.forEach((value, index) =>
-      expect(cells.eq(index).text().trim()).to.be.equal(value)
-    );
-    expect(this.$('.one-collapsible-list-item:last-child' +
-        ' .one-collapsible-list-item-header .oneicon.oneicon-checkbox-filled'))
-      .to.exist;
-  });
-
-  it('renders message about no reports', function () {
-    this.set('data', []);
-    this.render(hbs `{{space-cleaning-reports data=data _window=_window}}`);
-
-    expect(this.$('tbody tr')).to.have.length(1);
-    expect(this.$('tbody td')).to.contain('No reports available.');
-  });
-
-  it('allows to sort released size using raw bytes number', function (done) {
-    this.render(hbs `{{space-cleaning-reports data=data _window=_window}}`);
-
-    const releasedTh = this.$('th:nth-child(3)')[0];
-    click(releasedTh).then(() => {
-      click(releasedTh).then(() => {
-        expect(this.$('tbody tr:first-child td:nth-child(3)'))
-          .to.contain('1 MiB');
-        done();
+    return wait().then(() => {
+      expect(this.$('.one-collapsible-list .data-row')).to.have.length(2);
+      return click(
+        '.one-collapsible-list-item:last-child .one-collapsible-list-item-header'
+      ).then(() => {
+        const cells = this.$(
+          '.one-collapsible-list-item:last-child .one-collapsible-list-item-content .one-label'
+        );
+        const cellsValues = [
+          moment(data[1].startedAt).format(START_END_TIME_FORMAT),
+          moment(data[1].stoppedAt).format(START_END_TIME_FORMAT),
+          '999 KiB (out of 999 KiB)',
+          String(data[1].filesNumber),
+        ];
+        cellsValues.forEach((value, index) =>
+          expect(cells.eq(index).text().trim()).to.be.equal(value)
+        );
+        expect(this.$('.one-collapsible-list-item:last-child' +
+            ' .one-collapsible-list-item-header .oneicon.oneicon-checkbox-filled'
+          ))
+          .to.exist;
       });
     });
   });
+
+  it('renders message about no reports', function () {
+    this.on('fetchReports', () => resolve([]));
+    this.render(hbs `<div class="col-content">
+      {{space-cleaning-reports
+        spaceId="space_id1"
+        isCleanEnabled=true
+        fetchReports=(action "fetchReports")
+        _window=_window
+      }}
+    </div>`);
+
+    return wait().then(() => {
+      const $noDataRow = this.$('.no-data-row');
+      expect($noDataRow).to.exist;
+      expect($noDataRow).to.contain('No reports available.');
+    });
+  });
+
 });