+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-5114 Infinite-scroll list of auto-cleaning reports, file-popularity options and major refactor of space support views
 * VFS-5153 Added DNS autodetect option
 * VFS-5070 Added new conditions and using refactored API for auto-cleaning
 * VFS-4919 Added support for WebDAV storage